requirements.txt
requests
Pillow
RPi.GPIO
